Fix LoadError "cannot load such file rest-client"
Previously the rest-client dependency was declared in Gemfile but not
gemspec, so was not installed as a depencency when this gem was added
to an app's Gemfile. That caused LoadError to be raised when this gem's
methods called rest-client methods.

This moves all dependency declarations from Gemfile to gemspec, so they
will be recognized by Bundler and installed as expected, and to follow
convention.

This also deletes Gemfile.lock to avoid the confusion that could be
caused by updating it in this commit, and to follow convention.