Rename domain as subdomain
Previously one of the two main configuration requirements was
"domain". That term is generally used for a fully qualified domain
name including TLD. However, what was actually expected was only
the subdomain. Providing a full domain caused exceptions to be
raised, like OpenSSL::SSL::SSLError "certificate verify failed
(hostname mismatch)".

This renames "domain" as "subdomain" to improve clarity, and adds
backward-compatibility methods to warn that "domain" is deprecated.
This also adds a note in the README explaining what should be set.